fix: Handle test circular import issue
