+++ apps/editor.planx.uk/src/pages/FlowEditor/lib/store/editor.ts
@@ -41,6 +41,7 @@ import { getFlowDoc, subscribeToDoc } from "./../sharedb";
 import { type Store } from ".";
 import type { NavigationStore } from "./navigation";
 import type { SharedStore } from "./shared";
+import { stripTemplatedNodeProps } from "./utils/stripTemplatedNodeProps";
 
 let doc: Doc;
 
@@ -859,25 +860,3 @@ export const editorStore: StateCreator<
     return response;
   },
 });
-
-const TEMPLATED_NODE_PROPS = [
-  "isTemplatedNode",
-  "templatedNodeInstructions",
-  "areTemplatedNodeInstructionsRequired",
-] as const;
-
-/**
- * Strip template-authoring props from a node's data
- *
- * Used when pasting a node copied from a source template into a flow that isn't
- * itself a source template - those props only make sense there, so left behind
- * they'd falsely mark an ordinary node as templated
- */
-export const stripTemplatedNodeProps = (node: Store.Node): Store.Node => {
-  if (!node.data) return node;
-
-  const data = { ...node.data };
-  TEMPLATED_NODE_PROPS.forEach((prop) => delete data[prop]);
-
-  return { ...node, data };
-};
